Unnecessary scoring deleted and comments added
No need for many of the models' predictions to be scored using R^2, so I deleted those parts. I also added comments in the "smaller linear model" part that I forgot to add earlier, and fixed a bit of code redundancy around filling in missing values for the 'beds' and 'bedrooms' columns.